refactor: using utils module in oauth2 and changing variables name to standarize
